==> notebooks/03_database_modeling.ipynb <==
import sqlite3
import pandas as pd

# Charger le dataset nettoyé
df = pd.read_csv("../data/processed/listings_clean.csv")

# Créer / connecter à la base SQLite
conn = sqlite3.connect("../data/processed/airbnb_analysis.db")

# Exporter le dataframe vers SQLite
df.to_sql("listings", conn, if_exists="replace", index=False)

print("Base de données créée avec succès.")
print("Nombre de lignes importées :", len(df))

conn.close()


import sqlite3
import pandas as pd

conn = sqlite3.connect("../data/processed/airbnb_analysis.db")

# Voir les tables
tables = pd.read_sql("SELECT name FROM sqlite_master WHERE type='table';", conn)
tables



KPI 1 — Nombre d’annonces par ville

query = """
SELECT city, COUNT(*) AS nb_listings
FROM listings
GROUP BY city
ORDER BY nb_listings DESC;
"""
pd.read_sql(query, conn)


KPI 2 — Prix moyen + médiane par ville

query_mean = """
SELECT city,
       ROUND(AVG(price), 2) AS avg_price
FROM listings
GROUP BY city;
"""
mean_df = pd.read_sql(query_mean, conn)

query_median = """
WITH ranked AS (
  SELECT
    city,
    price,
    ROW_NUMBER() OVER (PARTITION BY city ORDER BY price) AS rn,
    COUNT(*) OVER (PARTITION BY city) AS cnt
  FROM listings
)
SELECT city,
       AVG(price) AS median_price
FROM ranked
WHERE rn IN ((cnt + 1) / 2, (cnt + 2) / 2)
GROUP BY city;
"""
median_df = pd.read_sql(query_median, conn)

mean_df.merge(median_df, on="city")


KPI 3 — Prix moyen par type de logement et par ville

query = """
SELECT city, room_type,
       ROUND(AVG(price), 2) AS avg_price,
       COUNT(*) AS nb_listings
FROM listings
GROUP BY city, room_type
ORDER BY city, avg_price DESC;
"""
pd.read_sql(query, conn)


KPI 4 — Top 10 quartiers les plus chers (par ville)

query = """
SELECT city, neighbourhood,
       ROUND(AVG(price), 2) AS avg_price,
       COUNT(*) AS nb_listings
FROM listings
GROUP BY city, neighbourhood
HAVING nb_listings >= 30
ORDER BY avg_price DESC
LIMIT 10;
"""
pd.read_sql(query, conn)


KPI 5 — Corrélation simple (prix vs disponibilité) en SQL

query = """
SELECT city,
       ROUND(AVG(price),2) AS avg_price,
       ROUND(AVG(availability_365),2) AS avg_availability,
       COUNT(*) AS n
FROM listings
GROUP BY city;
"""
pd.read_sql(query, conn)


conn.close()
print("Connexion SQLite fermée.")
